feat: Implement core database operations and file management
- Added Database class for managing collections and their operations.
- Introduced Indexation class for handling database creation and metadata.
- Created InMemoryCache for caching data in memory with TTL support.
- Developed FileManager and FolderManager for file and directory operations.
- Implemented utility classes for searching and sorting data.
- Updated server configuration to integrate new services and improve structure.
