+++ source/engine/Filesystem/FolderManager.ts
