+++ source/utility/HashMapSearch.utils.ts
